+++ src/XrmContext/CommandLine/CommandLineHelper.fs
@@ -13,6 +13,7 @@ module internal CommandLineHelper =
 
   let getListArg (args:Map<string,string>) arg transformer = 
     match Map.tryFind arg args with
+    | None -> None
     | Some value -> 
       value.Split([|','|], StringSplitOptions.RemoveEmptyEntries) 
       |> Array.map (fun s -> s.Trim())
@@ -21,7 +22,6 @@ module internal CommandLineHelper =
       |> function
       | arr when arr.Length > 0 -> Some arr
       | _ -> None
-    | None -> None
 
   let (|GetArgVal|_|) input = 
     let m = Regex("^/([^:]+):\"?(.*)\"?$").Match(input)
@@ -30,10 +30,13 @@ module internal CommandLineHelper =
 
   let handleArg expectedArgs parsedArgs k v =
     match Set.contains k expectedArgs with
-      | true -> Map.add k v parsedArgs
-      | false ->
-        printfn "Option '%s' not recognized." k
-        parsedArgs
+    | true -> Map.add k v parsedArgs
+    | false ->
+      printfn "Option '%s' not recognized." k
+      parsedArgs
+
+  let sharesElement set1 =
+    Set.intersect set1 >> Set.isEmpty >> not
 
   /// Helper function that recursively parses the arguments
   let rec parseCommandLineRec args expectedArgs parsedArgs =
@@ -51,15 +54,19 @@ module internal CommandLineHelper =
       handleArg expectedArgs args k ConfigurationManager.AppSettings.[k]
     ) parsedArgs
 
-
   /// Parses the given arguments against the expected arguments.
   let parseArgs argv expectedArgs =
     let argSet = expectedArgs |> List.map (fun a -> a.command.ToLower()) |> Set.ofList 
-    let argv = argv |> List.ofArray
+
+    let metaArgs, appArgs = argv |> Array.partition (fun a -> Args.useConfigArgs.Contains a)
+    let metaArgs = metaArgs |> Set.ofArray
+    let appArgs = appArgs |> List.ofArray
 
     let parsedArgs = 
-      parseConfigArgs argSet Map.empty
-      |> parseCommandLineRec argv argSet
+      Map.empty
+      |> if List.isEmpty appArgs || sharesElement metaArgs Args.useConfigArgs
+         then parseConfigArgs argSet else id
+      |> parseCommandLineRec appArgs argSet
   
     let missingArgs =
       expectedArgs
@@ -83,6 +90,12 @@ module internal CommandLineHelper =
           arg.command 
           arg.description 
           (if arg.required then " (required)" else ""))
+
+  let showDescription () =
+    printfn "[%s v.%s]" 
+      (Reflection.Assembly.GetExecutingAssembly().GetName().Name) 
+      System.AssemblyVersionInformation.Version
+    printfn ""
 
   let showUsage () =
     printfn "%s" Args.usageString